clear the MainActivity
