+++ app/src/main/res/layout/activity_current_position_of_driver.xml
@@ -5,11 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/fragment_container"/>
+    tools:context=".CurrentPositionOfDriver">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
